Deconvolves Thermogravimetric Data

This function deconvolves thermogravimetric data using a Fraser-Suzuki mixture model

deconvolve(process_object, lower_temp = 120, upper_temp = 700, seed = 1,
  n_curves = NULL, start_vec = NULL, lower_vec = NULL, upper_vec = NULL)

Arguments

process_object
process object obtained from process function

lower_temp
lower temperature bound to crop dataset, default to 120

upper_temp
upper temperature bound to crop dataset, default to 700

seed
random seed for nloptr optimiser

n_curves
number of curves optional specification

start_vec
vector of starting values for nls function. Only specify this vector if you have selected the number of curves in the n_curves parameter.

lower_vec
vector of lower bound values for nls. Only specify this vector if you have selected the number of curves in the n_curves parameter.

upper_vec
vector of upper bound values for nls. Only specify this vector if you have selected the number of curves in the n_curves parameter.

Value

decon list containing amended dataframe, bounds, model output, mass fractions

Examples

data(juncus)
tmp <- process(juncus, 'temp_C', 'mass_loss', 18.96)
output <- deconvolve(tmp)